[Task] Fix asynchronous cancel for Task
When GCancellable is called, context is first deleted then
callback to cleanup the task is called.

GUPnPServiceProxy retrieved in context, was saved in the Task User Data
Callback.

When the context was deleted, the proxy was unreferenced/deleted.
The use of the proxy in the task callback is then prohibited.

We add a weak pointer on the GUPnPServiceProxy so the pointer on
the proxy in the task user data is set to null when the object is deleted.

Signed-off-by: Ludovic Ferrandis <ludovic.ferrandis@intel.com>